refactor with ParquetFileFormat's mergeSchema
