+++ sql/hive/src/test/scala/org/apache/spark/sql/hive/orc/HiveOrcSourceSuite.scala
@@ -168,6 +168,6 @@ class HiveOrcSourceSuite extends OrcSuite with TestHiveSingleton {
   }
 
   test("SPARK-11412 read and merge orc schemas in parallel") {
-    testMergeSchemasInParallel(OrcFileOperator.singleFileSchemaReader)
+    testMergeSchemasInParallel(OrcFileOperator.readOrcSchemasInParallel)
   }
 }
